backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
pydantic==1.10.14
python-multipart==0.0.6
python-dotenv==1.0.0
pdfplumber==0.10.4
langchain==0.0.350
langchain-google-genai==0.0.8
google-generativeai==0.3.2
pytest==7.4.3
pytest-asyncio==0.21.1